// ssyweb/SSYNEW/src/app/models/pfCommissionConfig.model.ts
export class PFCommissionConfig {
    commissionConfigId: number;
    commissionCode: string;
    commissionRate: any;
    agentType: number;
    commissionType: string;
    effectiveDateFrom: Date;
    effectiveDateTo: Date;
    effectiveDateFromString: string;
    effectiveDateToString: string;
    status: number;
    arrearsRate: number;
    percentageOnNormalAmount: any;
    percentageOnArrearAmount: any;
    fixedNormalAmount: any;
    fixedArrearAmount: any;

}